Stop using ECDH_C in PSA built-in ECDH
Using ECP directly is less complex and reduces dependencies.
(This is similar to how FFDH uses BIGNUM directly rather than DHM_C.)

This saves 1kB of code size on a config similar to TF-M medium except
not using the p256-m driver. (That is, comment out
MBEDTLS_PSA_P256M_DRIVER_ENABLED in
configs/ext/tfm_mbedcrypto_config_profile_medium.h
before running mtest -S).

Before this commit: 54193
After this commit:  53117
Saved:               1076

Another benefit is that once ecdh.h is no longer part of the public
interface, we can completely remove ecdh.c and ecdh.h, and with them a
significant amount of complexity.

Currently this breaks support for Everest in PSA, as this relied on the
special dispatch implemented in ecdh.c. Support for Everest in PSA will
later be restored in the proper way: by making in a PSA driver.

Signed-off-by: Manuel Pégourié-Gonnard <manuel.pegourie-gonnard@arm.com>