Fix possible use-after-free in iop ECDH
This was never present in a released version, so it doesn't need a
ChangeLog entry or advisory, and can be fixed in public.

Introduced by https://github.com/Mbed-TLS/TF-PSA-Crypto/pull/578

We have a pre-existing test gap in all key agreement functions that we
don't test with a short input buffer for the peer's key. Closing this
gap for iop key agreement revealed the bug.

The next commit with close this gap for other key agreement functions.

Signed-off-by: Manuel Pégourié-Gonnard <manuel.pegourie-gonnard@arm.com>